Dev: remove chartData and merge methods into chart class
