+++ charts/sourcesTab/SourcesTab.tsx
@@ -3,12 +3,11 @@ import * as React from "react"
 import { computed } from "mobx"
 import { observer } from "mobx-react"
 import { Bounds } from "charts/utils/Bounds"
-import { ChartConfig } from "charts/core/ChartConfig"
+import { ChartConfig, SourceWithDimension } from "charts/core/ChartConfig"
 import * as Cookies from "js-cookie"
 import { ADMIN_BASE_URL } from "settings"
 import { faPencilAlt } from "@fortawesome/free-solid-svg-icons/faPencilAlt"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { SourceWithDimension } from "charts/core/ChartData"
 
 function formatText(s: string) {
     return linkify(s).replace(/(?:\r\n|\r|\n)/g, "<br/>")
